api-gateway: rename `controller` to `adapterin`
