+++ srv/api_gateway/adapterin/controller.module.go
@@ -1,4 +1,4 @@
-package controller
+package adapterin
 
 import (
 	"github.com/gin-gonic/gin"
