+++ srv/api_gateway/adapterin/http_handler.go
@@ -1,4 +1,4 @@
-package controller
+package adapterin
 
 import (
 	"context"
