+++ srv/api_gateway/adapterin/is_logged.go
@@ -1,4 +1,4 @@
-package controller
+package adapterin
 
 import (
 	"github.com/alimitedgroup/MVP/common/dto"
