+++ srv/api_gateway/adapterin/login_test.go
@@ -1,4 +1,4 @@
-package controller
+package adapterin
 
 import (
 	"encoding/json"
